// web/app/lib/gen/piano_user_list/v1/piano_user_list_pb.ts
export enum PianoListKind {
  /**
   * @generated from enum value: PIANO_LIST_KIND_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: PIANO_LIST_KIND_WISHLIST = 1;
   */
  WISHLIST = 1,

  /**
   * @generated from enum value: PIANO_LIST_KIND_VISITED = 2;
   */
  VISITED = 2,

  /**
   * @generated from enum value: PIANO_LIST_KIND_FAVORITE = 3;
   */
  FAVORITE = 3,
}
// ...
export class ListUserListPianosRequest extends Message<ListUserListPianosRequest> {
  /**
   * "users/{custom_id}"
   *
   * @generated from field: string parent = 1;
   */
  parent = "";

  /**
   * @generated from field: piano_user_list.v1.PianoListKind list_kind = 2;
   */
  listKind = PianoListKind.UNSPECIFIED;

  /**
   * @generated from field: optional int32 page_size = 3;
   */
  pageSize?: number;

  /**
   * @generated from field: optional string page_token = 4;
   */
  pageToken?: string;

  constructor(data?: PartialMessage<ListUserListPianosRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "piano_user_list.v1.ListUserListPianosRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "parent", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "list_kind", kind: "enum", T: proto3.getEnumType(PianoListKind) },
    { no: 3, name: "page_size", kind: "scalar", T: 5 /* ScalarType.INT32 */, opt: true },
    { no: 4, name: "page_token", kind: "scalar", T: 9 /* ScalarType.STRING */, opt: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ListUserListPianosRequest {
    return new ListUserListPianosRequest().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ListUserListPianosRequest {
    return new ListUserListPianosRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ListUserListPianosRequest {
    return new ListUserListPianosRequest().fromJsonString(jsonString, options);
  }

  static equals(a: ListUserListPianosRequest | PlainMessage<ListUserListPianosRequest> | undefined, b: ListUserListPianosRequest | PlainMessage<ListUserListPianosRequest> | undefined): boolean {
    return proto3.util.equals(ListUserListPianosRequest, a, b);
  }
}